+++ lib/models/section.dart
@@ -72,7 +72,8 @@ class Section extends ChangeNotifier {
     }
 
     for (final original in originalItems!) {
-      if (!items!.contains(original)) {
+      if (!items!.contains(original) &&
+          (original.image as String).contains('firebase')) {
         try {
           final ref = await storage.refFromURL(original.image as String);
           await ref.delete();
@@ -90,10 +91,12 @@ class Section extends ChangeNotifier {
   Future<void> delete() async {
     await firestoreRef.delete();
     for (final item in items!) {
-      try {
-        final ref = await storage.refFromURL(item.image as String);
-        await ref.delete();
-      } catch (e) {}
+      if ((item.image as String).contains('firebase')) {
+        try {
+          final ref = await storage.refFromURL(item.image as String);
+          await ref.delete();
+        } catch (e) {}
+      }
     }
   }
